smart_backend/requirements.txt
asgiref==3.11.1
attrs==25.4.0
Django==5.2
django-cors-headers==4.4.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
